moving rdmArticleData to a shared util file
